# Conflicts:
#	engine-api/pom.xml
#	engine-api/src/main/java/com/dtstack/engine/api/ApiUrlUtil.java
#	engine-api/src/main/java/com/dtstack/engine/api/dto/UserDTO.java
#	engine-dao/src/main/resources/sqlmap/scheduleJob-mapper.xml
#	engine-master/src/main/java/com/dtstack/engine/master/executor/JobExecutorTrigger.java
#	engine-master/src/main/java/com/dtstack/engine/master/impl/ScheduleJobService.java

+++ engine-plugins/flink/flink180-HW/src/main/java/org/apache/flink/optimizer/plantranslate/JobGraphGenerator.java
@@ -302,7 +302,9 @@ public class JobGraphGenerator implements Visitor<PlanNode> {
 				throw new FlinkRuntimeException("Could not compress distributed-cache artifacts.", ioe);
 			} finally {
 				try {
-					Files.deleteIfExists(tmpDir);
+					if (tmpDir != null) {
+						Files.deleteIfExists(tmpDir);
+					}
 				} catch (IOException e) {
 					e.printStackTrace();
 				}
